requirements.txt
streamlit
rdkit-pypi
pandas
matplotlib
numpy
openpyxl
pubchempy

